+++ app.js
@@ -843,6 +843,9 @@ import { renderers as RV1490 } from './views/group-v1490.js';
 import { renderers as RV1491 } from './views/group-v1491.js';
 import { renderers as RV1492 } from './views/group-v1492.js';
 import { renderers as RV1493 } from './views/group-v1493.js';
+import { renderers as RV1494 } from './views/group-v1494.js';
+import { renderers as RV1495 } from './views/group-v1495.js';
+import { renderers as RV1496 } from './views/group-v1496.js';
 import { renderers as RV1396 } from './views/group-v1396.js';
 import { renderers as RV164 } from './views/group-v164.js';
 import { renderers as RV165 } from './views/group-v165.js';
@@ -931,7 +934,7 @@ const RENDERERS = { ...RA, ...RB, ...RC, ...RE, ...RF, ...RG, ...RH, ...RI, ...R
   ...RV902,
   ...RV903,
   ...RV905,
-  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV1418, ...RV1419, ...RV1420, ...RV1421, ...RV1422, ...RV1423, ...RV1424, ...RV1425, ...RV1426, ...RV1427, ...RV1428, ...RV1429, ...RV1430, ...RV1431, ...RV1433, ...RV1434, ...RV1435, ...RV1436, ...RV1438, ...RV1439, ...RV1440, ...RV1441, ...RV1442, ...RV1443, ...RV1444, ...RV1445, ...RV1446, ...RV1447, ...RV1448, ...RV1449, ...RV1450, ...RV1451, ...RV1452, ...RV1453, ...RV1454, ...RV1455, ...RV1468, ...RV1469, ...RV1470, ...RV1471, ...RV1472, ...RV1473, ...RV1475, ...RV1476, ...RV1479, ...RV1480, ...RV1481, ...RV1482, ...RV1483, ...RV1484, ...RV1485, ...RV1486, ...RV1487, ...RV1488, ...RV1489, ...RV1490, ...RV1491, ...RV1492, ...RV1493, ...RV63, ...RPALINT };
+  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV1418, ...RV1419, ...RV1420, ...RV1421, ...RV1422, ...RV1423, ...RV1424, ...RV1425, ...RV1426, ...RV1427, ...RV1428, ...RV1429, ...RV1430, ...RV1431, ...RV1433, ...RV1434, ...RV1435, ...RV1436, ...RV1438, ...RV1439, ...RV1440, ...RV1441, ...RV1442, ...RV1443, ...RV1444, ...RV1445, ...RV1446, ...RV1447, ...RV1448, ...RV1449, ...RV1450, ...RV1451, ...RV1452, ...RV1453, ...RV1454, ...RV1455, ...RV1468, ...RV1469, ...RV1470, ...RV1471, ...RV1472, ...RV1473, ...RV1475, ...RV1476, ...RV1479, ...RV1480, ...RV1481, ...RV1482, ...RV1483, ...RV1484, ...RV1485, ...RV1486, ...RV1487, ...RV1488, ...RV1489, ...RV1490, ...RV1491, ...RV1492, ...RV1493, ...RV1494, ...RV1495, ...RV1496, ...RV63, ...RPALINT };
 
 // ----- Utility registry ----------------------------------------------------
 // Source of truth for routes, names, group, audiences, and clinical flag.
@@ -2219,6 +2222,9 @@ const UTILITIES = [
   { id: 'pufa-index', name: 'PUFA Index (Untreated Caries)', group: 'G', audiences: ['clinicians',  'educators'], clinical: true },
   { id: 'bolton-ratio', name: 'Bolton Tooth-Size Ratio', group: 'G', audiences: ['clinicians',  'educators'], clinical: true },
   { id: 'eichner-index', name: 'Eichner Index (Occlusal Support)', group: 'G', audiences: ['clinicians',  'educators'], clinical: true },
+  { id: 'kotlow-ankyloglossia', name: 'Kotlow Ankyloglossia Class', group: 'G', audiences: ['clinicians',  'educators'], clinical: true },
+  { id: 'ohat-oral-health', name: 'Oral Health Assessment Tool (OHAT)', group: 'G', audiences: ['clinicians',  'educators'], clinical: true },
+  { id: 'turesky-plaque', name: 'Turesky Plaque Index (TQHPI)', group: 'G', audiences: ['clinicians',  'educators'], clinical: true },
   { id: 'pederson-difficulty', name: 'Pederson Difficulty (Third Molar)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'ellis-tooth-fracture', name: 'Ellis Dental Fracture Class', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'kennedy-edentulous', name: 'Kennedy Classification (RPD)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },